refactor: rename the fakegithub to fakegitprovider package to reflect its purpose
